update tests for conditional rendering page logic
