+++ stories/signup/Signup.stories.tsx
@@ -1,10 +1,9 @@
 import * as React from 'react'
 import Signup from '../../pages/signup'
-import { Props } from '../../@types/signup'
 
 export default {
   component: Signup,
   title: 'Pages/signup'
 }
 
-export const _Signup: React.FC<Props> = props => <Signup {...props} />
+export const _Signup: React.FC = () => <Signup />
