--- stories/signup/Success.stories.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react'
-import Success from '../../pages/signup/success'
-
-export default {
-  component: Success,
-  title: 'Pages/signup'
-}
-
-export const _Success: React.FC = () => <Success />
